fix opponent factor compitation

+++ Source/opponent.hpp
@@ -137,7 +137,7 @@ void init(){
           for (auto const& it : ratings){
              if ( (oppName.find(str_tolower(it.first)) != std::string::npos) || (str_tolower(it.first).find(oppName) != std::string::npos) ){
                 oppRating = it.second;
-                DynamicConfig::ratingFactor = std::tanh((myRating/oppRating)-1)*20.;
+                DynamicConfig::ratingFactor = std::tanh((myRating/double(oppRating))-1)*20.;
                 break;
              }
           }
